Client can now display a list of trip notes and add a new note
